чекбокс больше не сохраняет состояние на странице сохраненные фильмы, кнопка сохранения в форме профиль заблокирована на время отправки запроса
